add routing and class active for links
