+++ src/index.scss
@@ -9,6 +9,18 @@ a {
   text-decoration: none;
 }
 
+.active {
+  background-color: #FBFBFB;
+  border-left: 5px solid #007BFF;
+  color: #007BFF;
+
+  svg {
+    path {
+      fill: #007BFF;
+    }
+  }
+}
+
 .btn {
   width: 220px;
   height: 58px;
@@ -80,11 +92,6 @@ a {
         border: 1px solid #BECAD6;
         border-radius: 6px;
         margin-bottom: 15px;
-      }
-
-      .login__btns {
-        display: flex;
-        justify-content: space-between;
       }
     }
   }
